rpk: move cmd/wasm to its own package
