+++ src/go/rpk/pkg/cli/cmd/wasm/remove.go
@@ -10,7 +10,7 @@ import (
 	"github.com/spf13/cobra"
 )
 
-func NewRemoveCommand(fs afero.Fs) *cobra.Command {
+func newRemoveCommand(fs afero.Fs) *cobra.Command {
 	var coprocType string
 
 	cmd := &cobra.Command{
